README.md,scan_open_ports.go: Implementation first version
scan_open_ports.go: This loops through all the ports from 1-32000 for a
given host via telnet with TCP to see if connection exists.
This is concurrent and limited by the limit var which is derived
from multiples of number of cores of the machine.

README.md: Added help, usage, simple performance tests